scripts/PowerPoint Slide Inventory Script/requirements.txt
python-pptx>=1.0.0
pyyaml>=6.0
Pillow>=10.0

# Phase 3 — vision critique pipeline (optional)
# Install before running render_from_yaml.py --critique
openai>=1.0           # vision model API (needs OPENAI_API_KEY env var)
pywin32>=306          # for render_slide_previews.py on Windows (PowerPoint COM)
                      # after install: python -m pywin32_postinstall -install

